Rearrange functions to split base CloudProvider
Split base class from its implementations.
Add mock tests for basic Dropbox functions.
