+++ hpcflow/__init__.py
@@ -1,1 +1,12 @@
 """`hpcflow.__init__.py`"""
+
+from hpcflow._version import __version__
+from hpcflow.api import (
+    make_workflow,
+    submit_workflow,
+    clean,
+    get_stats,
+    save_stats,
+    kill,
+    cloud_connect,
+)
