+++ hpcflow/api.py
@@ -12,12 +12,12 @@ import json
 from beautifultable import BeautifulTable
 
 from hpcflow.archive.archive import CloudProviderType
+import hpcflow.archive.cloud.dropbox_cp as dropbox
 from hpcflow.config import Config
 from hpcflow.init_db import init_db
 from hpcflow.models import Workflow, CommandGroupSubmission
 from hpcflow.profiles import parse_job_profiles, prepare_workflow_dict
 from hpcflow.project import Project
-from hpcflow.archive.cloud.cloud import CloudProvider, DropboxCloudProvider, get_token
 
 
 def make_workflow(dir_path=None, profile_list=None, json_file=None, json_str=None,
@@ -453,5 +453,5 @@ def cloud_connect(provider, config_dir=None):
             pass
     if not good:
         print('Getting new cloud token.')
-        token = get_token()
+        token = dropbox.get_token()
         update_config(token_key, token, config_dir=config_dir)
